+++ project/src/components/elements/favorites-empty/favorites-empty.jsx
