+++ project/src/components/elements/offer-card/offer-card.jsx
