+++ project/src/components/elements/room-gallery/room-gallery.jsx
