+++ project/src/components/pages/not-found-screen/not-found-screen.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import Header from '../../header/header';
+import Header from '../../elements/header/header';
 
 function NotFoundScreen () {
   return (
