--- project/src/components/sort-item/sort-item.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { SortType } from '../../const';
-
-function SortItem({type, isActive, onClick}) {
-  return (
-    <li
-      className={`places__option ${isActive ? 'places__option--active' : ''}`}
-      tabIndex="0"
-      onClick={onClick}
-    >
-      {SortType[type].text}
-    </li>
-  );
-}
-
-SortItem.propTypes = {
-  type: PropTypes.string.isRequired,
-  isActive: PropTypes.bool.isRequired,
-  onClick: PropTypes.func.isRequired,
-};
-
-export default SortItem;
